Implement filters for products. Implements gallery for homepage. Several corrections.

+++ product/src/main/java/com/satgy/embudi/product/service/FileServiceImp.java
@@ -5,7 +5,6 @@ import com.satgy.embudi.product.repository.FileRepositoryI;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -14,9 +13,17 @@ public class FileServiceImp implements FileServiceI {
     @Autowired
     private FileRepositoryI repo;
 
+    @Autowired
+    private FileStorageService fileStorageService;
+
     @Override
     public Optional<File> findById(Long id) {
         return repo.findById(id);
+    }
+
+    @Override
+    public Optional<File> findByProductResource(Long productResourceId) {
+        return repo.findByProductResource(productResourceId);
     }
 
     @Override
@@ -25,17 +32,21 @@ public class FileServiceImp implements FileServiceI {
     }
 
     @Override
-    public File create(File country) {
-        return repo.save(country);
+    public File create(File file) {
+        return repo.save(file);
     }
 
     @Override
-    public File update(File country) {
-        return repo.save(country);
+    public File update(File file) {
+        return repo.save(file);
     }
 
     @Override
     public void delete(Long id) {
-        repo.deleteById(id);
+        Optional<File> of = repo.findById(id);
+        if (of.isPresent()){
+            fileStorageService.deleteFileAsResource(of.get().getName());
+            repo.deleteById(id);
+        }
     }
 }
